docker/inference/requirements.txt
flask==3.0.3
tensorflow>=2.15.0
pillow>=10.0.0
numpy>=1.24.0

